Use service wrapper to facilitate using dashing-go on windows
Introduces bitbucket.org/kardianos/service to create a windows
service, as well as support for installing as a service on
mac os x and linux. Also ensures working directory when the
server starts is correct (this is important when starting
as a windows service).

Uses github.com/patdowney/dashing-go for the moment.